visp_lilo: update sensor pipeline, formats, and cam_device interfaces
- update top-level and visp build rules for new ISP integration and
  module layout
- consolidate sensor/common headers and refresh v4l2 extensions
- drop legacy ISP ctrl versions; keep active v5.2/v2 paths
- switch default sensor calibration to JSON and adjust procfs fields
- add sink-format-based sensor mode matching and stricter pipeline setup
- extend cam_device API structs (CCM, metadata, stride/offset, MCM cfg)
- align mbox interfaces and cam_device sensor mapping for updated targets
- additional fixes: harden sensor control copies, guard format.reserved,
  fix compat-flag leak, and make procfs string updates safe

+++ visp_lilo/visp_event.h
@@ -75,7 +75,7 @@ enum visp_vevent_id {
 };
 
 struct visp_plane {
-	uint32_t dma_addr;
+	uint64_t dma_addr;
 	uint32_t size;
 };
